StepDir: Added configurable calculation precision
The precision of the calculation is always kept equal to
the frequency. The StepDir frequency is dictated by the worst case
calculation time of the interrupt. Previously any board with a single
motor was still limited by the worst-case of a board with two motors.
By having a configurable precision these boards can now double their
frequency while still keeping frequency == precision.

Additionally, due to us now checking StallGuard status in the interrupt
in the EvalSystem we started being too slow for our old precision of 17
bits. Instead of cutting that precision (and frequency) in half we can
now adjust by less than a factor of two.