StepDir: Fixed occasional jitter around the target position
Sometimes when moving to a position the StepDir generator would jitter
around the target position (+-1 steps). This caused vibration of the
motor and prevented current standstill reduction from activating.

The accelerationSteps variable counts the amount of steps needed for
decelerating to 0 velocity. Due to rounding errors from the accumulator
logic, this count is not always accurate. If this count turns negative,
a moving ramp will not stop moving before it has already surpassed its
target position. Additionally this caused the homing procedure meant for
reaching a barely missed target position to instead switch over to the
normal driving mode. The combination of these issues caused the
oscillation around the target position.