General
Renamed FIELD_UPDATE to FIELD_WRITE for every IC, as we are actually
writing to some field. The first implementation of FIELD_WRITE was
plainly writing without previous readout and modification. Since this
macro is never used and also not implemented in most chips, just using
the FIELD_WRITE instead of FIELD_UPDATE suffices.